--- a/src/llama-context.h
+++ b/src/llama-context.h
@@ -198,9 +198,9 @@ public:
 
     // reserve a graph with a dummy ubatch of the specified size
     ggml_cgraph * graph_reserve(uint32_t n_tokens, uint32_t n_seqs, uint32_t n_outputs, const llama_memory_context_i * mctx, bool split_only = false);
-#ifdef GGML_PERF
+#if defined(GGML_PERF) || defined(GGML_PERF_DETAIL)
     struct ggml_perf_totals perf_totals[GGML_OP_COUNT] = {};  // add this to llama_context
-#endif
+#endif /* GGML_PERF  || GGML_PERF_DETAIL */
 
 private:
     llm_graph_params graph_params(